+++ packages/db/src/schema/index.ts
@@ -20,3 +20,4 @@ export * from './returns';
 export * from './inter_branch_transfers';
 export * from './inter_branch_payables';
 export * from './cash_flow';
+export * from './customer_portal';